// Clocktower-UI/src/openApi/clocktowerServer.d.ts
export interface paths {
// ...
    "/api/games/{gameId}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get the game state by id */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    gameId: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Clocktower.Server.Data.GameState"];
                    };
                };
                /** @description Not Found */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": string;
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/games": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Gets all games */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Clocktower.Server.Data.GameState"][];
                    };
                };
                /** @description Not Found */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": string;
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/games/load": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** Loads dummy data from saved json file */
        post: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": string;
                    };
                };
                /** @description Bad Request */
                400: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": string;
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/games/{gameId}/start": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        /** Starts new game state for id */
        post: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    gameId: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Created */
                201: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["Clocktower.Server.Data.GameState"];
                    };
                };
                /** @description Bad Request */
                400: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/problem+json": components["schemas"]["Microsoft.AspNetCore.Http.HttpValidationProblemDetails"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}

export type webhooks = Record<string, never>;

export interface components {
    schemas: {
        "Clocktower.Server.Data.GameState": {
            id?: string | null;
            players?: components["schemas"]["Clocktower.Server.Data.Types.Player"][] | null;
            /** Format: int32 */
            maxPlayers?: number;
            readonly isFull?: boolean;
        };
        /** @enum {string} */
        "Clocktower.Server.Data.Types.Enum.Edition": "TroubleBrewing" | "SectsAndViolets" | "BadMoonRising" | "Experimental";
        /** @enum {string} */
        "Clocktower.Server.Data.Types.Enum.RoleType": "Townsfolk" | "Outsider" | "Minion" | "Demon" | "Traveller";
        "Clocktower.Server.Data.Types.Player": {
            /** Format: int32 */
            id?: number;
            name?: string | null;
        };
        "Clocktower.Server.Data.Types.Role.Role": {
            name?: string | null;
            description?: string | null;
            type?: components["schemas"]["Clocktower.Server.Data.Types.Enum.RoleType"];
            edition?: components["schemas"]["Clocktower.Server.Data.Types.Enum.Edition"];
        };
        "Clocktower.Server.Discord.Endpoints.CheckGuild.CheckGuildResponse": {
            valid?: boolean;
            name?: string | null;
            message?: string | null;
        };
        "Clocktower.Server.Discord.Endpoints.TownStatus.TownExistsResponse": {
            exists?: boolean;
            message?: string | null;
        };
        "Clocktower.Server.Discord.Services.ChannelOccupants": {
            channel?: components["schemas"]["Clocktower.Server.Discord.Services.MiniChannel"];
            occupants?: components["schemas"]["Clocktower.Server.Discord.Services.MiniUser"][] | null;
        };
        "Clocktower.Server.Discord.Services.MiniCategory": {
            /** Format: int64 */
            id?: number;
            name?: string | null;
            channels?: components["schemas"]["Clocktower.Server.Discord.Services.ChannelOccupants"][] | null;
        };
        "Clocktower.Server.Discord.Services.MiniChannel": {
            /** Format: int64 */
            id?: number;
            name?: string | null;
        };
        "Clocktower.Server.Discord.Services.MiniUser": {
            /** Format: int64 */
            id?: number;
            name?: string | null;
        };
        "Clocktower.Server.Discord.Services.TownOccupants": {
            channelCategories?: components["schemas"]["Clocktower.Server.Discord.Services.MiniCategory"][] | null;
            /** Format: int32 */
            readonly userCount?: number;
        };
        "Clocktower.Server.Roles.Endpoints.GetRoles.Response": {
            roles?: components["schemas"]["Clocktower.Server.Data.Types.Role.Role"][] | null;
        };
        "Microsoft.AspNetCore.Http.HttpValidationProblemDetails": {
            type?: string | null;
            title?: string | null;
            /** Format: int32 */
            status?: number | null;
            detail?: string | null;
            instance?: string | null;
            errors?: {
                [key: string]: string[];
            } | null;
        } & {
            [key: string]: unknown;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}